Primitives: introduce `SignalHandler`
Add a platform primitive for handling signals on Unix platforms. Signals
are complicated in that the signal context cannot perform a majority of
actions. This adds a core primitive to install and uninstall signal
handlers. It is not fully signal safe though as `Task { ... }` may
result in `malloc` which is not signal safe.

This provides a starting point for signal handling, which is required
for restoring the terminal IO state and monitoring terminal resizes.